+++ app/src/main/java/com/vdemelo/starwarswiki/di/PresentationModule.kt
@@ -1,6 +1,5 @@
 package com.vdemelo.starwarswiki.di
 
-import com.vdemelo.starwarswiki.ui.HomeViewModel
 import com.vdemelo.starwarswiki.ui.screens.species.details.SpeciesDetailsViewModel
 import com.vdemelo.starwarswiki.ui.screens.species.list.SpeciesListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -9,5 +8,4 @@ import org.koin.dsl.module
 val presentationModule = module {
     viewModel { SpeciesListViewModel(get()) }
     viewModel { SpeciesDetailsViewModel(get()) }
-    viewModel { HomeViewModel(get()) }
 }
