+++ app/src/main/java/com/vdemelo/starwarswiki/ui/screens/base/UnexpectedErrorGoBackScreen.kt
@@ -1,0 +1,8 @@
+package com.vdemelo.starwarswiki.ui.screens.base
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun UnexpectedErrorGoBackScreen() {
+    //TODO criar telinha top
+}
